submodules: lvgl: Pull feat/demo-high-res-pr
Pull the feat/demo-high-res-pr branch from a developer's fork [1]. Work
is still on-going on that branch, hence it is not merged to master. But
I need to start working on it in parallel, so pull the branch in for
now. As changes are made to the upstream branch, I will keep pulling
them here as well, until they're merged.

[1] https://github.com/liamhowatt/lvgl

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>